TextToEmoji command crash fix, we can choose if MessageEditor deletes the message

+++ src/main/java/org/sdd/shenron/util/MessageEditor.java
@@ -19,10 +19,19 @@ public final class MessageEditor
 
     public static void edit(User user, Conversation conversation, Message message, String content) throws Exception
     {
+        edit(user, conversation, message, content, true);
+    }
+
+    public static void edit(User user, Conversation conversation, Message message, String content, boolean delete) throws Exception
+    {
         if (user instanceof DiscordUser)
         {
             edit((DiscordUser) user, (DiscordConversation) conversation, content);
-            message.delete();
+
+            if (delete)
+            {
+                message.delete();
+            }
         }
         else
         {
